fix(server): split clustering test modules out of oversized relay files
Qodo compliance rule 513872: route_bridge.rs, relay.rs, and
ordered_relay.rs mixed production logic with large in-file test
modules. Move each cfg(test) module to a child file
(ordered_relay/{invariant_tests,tests}.rs, relay/tests.rs,
route_bridge/tests.rs), preserving the debug_assertions gate on the
ordered-relay behavior matrix and fixing the include_str! path in the
dispatch-span guard test. No production code moved or changed.
